+++ FittsInstanceDwell.pde
@@ -7,7 +7,7 @@ class FittsInstanceDwell extends FittsInstance {
   }
 
 
-  public void update(float frameTimeMillis, HashMap<Action, Float> readings) {
+  public void update(long frameTimeMillis, HashMap<Action, Float> readings) {
     super.update(frameTimeMillis, readings);
     if (currentState.isCursorInTarget())
       currentDwellTimeMillis += frameTimeMillis;
